ref: move db into independent module
The declaration of entities and JPA repositories has no ties with the
application server. It is also cleaner to share these resources when the
module can be compiled as a library and not an executable jar (which
spring's repackage does).

Therefore all DB related classes have been moved into their own module,
this makes it easy to use them for dev tools without having to pull in
the whole application.
