+++ Dockerfile
@@ -3,7 +3,7 @@ FROM maven:3.9.6-eclipse-temurin-21-alpine AS build
 COPY / /src
 
 RUN apk add --update npm;
-RUN mvn -Dmaven.test.skip -f /src/felf/pom.xml clean package
+RUN mvn -Dmaven.test.skip -f /src/pom.xml clean package
 
 FROM eclipse-temurin:21-jre-alpine
 COPY --from=build /src/felf/target/felf.jar /opt/app/felf.jar
