+++ felf-db/src/main/java/com/zoftko/felf/db/dao/InstallationRepository.java
@@ -1,6 +1,6 @@
-package com.zoftko.felf.dao;
+package com.zoftko.felf.db.dao;
 
-import com.zoftko.felf.entities.Installation;
+import com.zoftko.felf.db.entities.Installation;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
