+++ felf-db/src/main/java/com/zoftko/felf/db/dao/ProjectRepository.java
@@ -1,6 +1,6 @@
-package com.zoftko.felf.dao;
+package com.zoftko.felf.db.dao;
 
-import com.zoftko.felf.entities.Project;
+import com.zoftko.felf.db.entities.Project;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
