+++ felf/src/test/java/com/zoftko/felf/services/WebhookServiceTests.java
@@ -7,10 +7,10 @@ import static org.mockito.Mockito.*;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.zoftko.felf.dao.InstallationRepository;
-import com.zoftko.felf.dao.ProjectRepository;
-import com.zoftko.felf.entities.Installation;
-import com.zoftko.felf.entities.Project;
+import com.zoftko.felf.db.dao.InstallationRepository;
+import com.zoftko.felf.db.dao.ProjectRepository;
+import com.zoftko.felf.db.entities.Installation;
+import com.zoftko.felf.db.entities.Project;
 import jakarta.persistence.EntityManager;
 import org.json.JSONException;
 import org.json.JSONObject;
