+++ pom.xml
@@ -17,14 +17,43 @@
 
     <properties>
         <java.version>21</java.version>
+        <spring-shell.version>3.1.4</spring-shell.version>
+        <jacoco.version>0.8.11</jacoco.version>
         <sonar.organization>zoftko</sonar.organization>
         <sonar.host.url>https://sonarcloud.io</sonar.host.url>
+        <sonar.sources>.</sonar.sources>
+        <sonar.inclusions>src/main/**</sonar.inclusions>
     </properties>
 
     <modules>
+        <module>felf-db</module>
         <module>felf</module>
         <module>fakelf</module>
+        <module>felf-coverage</module>
     </modules>
+
+    <profiles>
+        <profile>
+            <id>coverage</id>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.jacoco</groupId>
+                        <artifactId>jacoco-maven-plugin</artifactId>
+                        <version>${jacoco.version}</version>
+                        <executions>
+                            <execution>
+                                <id>prepare-agent</id>
+                                <goals>
+                                    <goal>prepare-agent</goal>
+                                </goals>
+                            </execution>
+                        </executions>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+    </profiles>
 
     <build>
         <plugins>
